11.30

+++ template/组合数.cpp
@@ -7,7 +7,7 @@ LL frac[N];  // 阶乘
 
 void init() {
     frac[0] = frac[1] = 1;
-    for (int i = 2; i < N; ++i) frac[i] = frac[i - 1] * i % mod;
+    for (LL i = 2; i < N; ++i) frac[i] = frac[i - 1] * i % mod;
 }
 
 LL fpow(LL x, LL exp) {  // 快速幂
@@ -20,7 +20,7 @@ LL fpow(LL x, LL exp) {  // 快速幂
 }
 
 /* C(n, k) means select k from n. */
-int C(int n, int k) {
+LL C(LL n, LL k) {
     LL upper = frac[n];
     LL lower = (fpow(frac[k], mod - 2) * fpow(frac[n - k], mod - 2)) % mod;  // 逆元
